ngr: Fix Gradle test runner by conditionally invoking `gradle wrapper`
In order to support projects not including a Gradle wrapper, invoking
`gradle wrapper` was sufficient to create one.

However, this leads to problems on other projects which include a
wrapper already:

> Could not set unknown property 'baseName' for task ':jar' of type org.gradle.api.tasks.bundling.Jar.
>
> -- https://github.com/crate/crate-sample-apps/actions/runs/6974312523/job/18979796291#step:5:40

This patch intends to improve the situation by only installing a wrapper
when ngr does not find a `gradlew` file in the directory.